Cho 2 nhan vat danh nhau

+++ systems/AttackSystem.cpp
@@ -18,32 +18,29 @@ void AttackSystem::begin(){}
 
 void AttackSystem::processEntity(artemis::Entity &e){
 	AttackComponent* attack = attackMapper.get(e);
+	attack->timeAlive += world->getDelta();
 	if (attack->whoAttack == R::CharacterType::NONAME){ return; }
-	if (attack->powerOfAttack<0 || attack->expire){ world->getEntityManager()->remove(e);  return; }
+	if (attack->timeAlive > attack->maxTimeAlive) { world->getEntityManager()->remove(e);   return; }
+	if (attack->powerOfAttack < 0 || attack->expire){ world->getEntityManager()->remove(e);  return; }
+	if (attack->timeAlive < attack->timeAttack) return;
 	if (attack->whoAttack == R::CharacterType::GOKU){
-		CCLOG("Check Attack");
 		// Lấy ra danh sách entity khác rồi xử lý chúng.
 		artemis::ImmutableBag<artemis::Entity*> *bag = world->getGroupManager()->getEntities("enemy");
-		if (bag->isEmpty()) return;
+		if (bag->isEmpty()) { world->getEntityManager()->remove(e);   return; }
 		for (int i = 0; i < bag->getCount(); i++){
 			artemis::Entity* entity = bag->get(i);
-			if (attackToEntity(e, *entity)) { 
+			if (attackToEntity(e, *entity)) {
 				RenderComponent* render = (RenderComponent*)e.getComponent<RenderComponent>();
-				if (render){
-				render->node->removeFromParentAndCleanup(true);
-				}
-				world->getEntityManager()->remove(e); 
-				return;
+				if (render)
+					render->node->removeFromParentAndCleanup(true);
 			}
 		}
 	}
 	else{
 		artemis::Entity &goku = world->getTagManager()->getEntity("goku");
-		if (attackToEntity(e, goku)){
-			world->getEntityManager()->remove(e);
-			return;
-		}
+		attackToEntity(e, goku);
 	}
+	world->getEntityManager()->remove(e);
 }
 void AttackSystem::end(){}
 
@@ -56,7 +53,7 @@ bool AttackSystem::attackToEntity(artemis::Entity& attackEntity, artemis::Entity
 	AttackComponent* defense = attackMapper.get(entity);
 	PosComponent* defensePosition = posMapper.get(entity);
 	BoundComponent* defenseBound = boundMapper.get(entity);
-	CharacterInfoComponent* defenseInfo = (CharacterInfoComponent*) entity.getComponent<CharacterInfoComponent>();
+	CharacterInfoComponent* defenseInfo = (CharacterInfoComponent*)entity.getComponent<CharacterInfoComponent>();
 
 	Rect attackRect = Rect(attackPosition->x + attackBound->x1, attackPosition->y + attackBound->y1, attackBound->getWidth(), attackBound->getHeight());
 	Rect defenseRect = Rect(defensePosition->x + defenseBound->x1, defensePosition->y + defenseBound->y1, defenseBound->getWidth(), defenseBound->getHeight());
@@ -64,16 +61,14 @@ bool AttackSystem::attackToEntity(artemis::Entity& attackEntity, artemis::Entity
 	if (attackRect.intersectsRect(defenseRect)){
 		StateComponent* defenseState = (StateComponent*)entity.getComponent<StateComponent>();
 		bool isLeft = defensePosition->x < attackPosition->x;
-		defenseState->customAnimation = true;
-		defenseState->animations.clear();
-		defenseState->animations.push_back("Trungdon");
-		defenseState->setState(R::CharacterState::TRUNG_DON);
+		defenseState->setState(R::CharacterState::DEFENSE);
+		defenseState->defense = R::Defense::TRUNG_DON;
 		defenseState->direction = isLeft ? R::Direction::LEFT : R::Direction::RIGHT;
 		defenseInfo->blood -= attack->powerOfAttack;
 		attack->expire = true;
 		return true;
 	}
-	
+
 
 	return false;
 }
